requirements.txt
flask
scapy
yara-python
hashlib
